requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
setuptools>=65.0.0
drf-spectacular==0.27.2

